requirements.txt
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.5
requests==2.23.0
whitenoise==5.2.0
